requirements.txt
nonebot2[fastapi]>=2.3.0
nonebot-adapter-onebot>=2.4.0
nonebot-plugin-apscheduler>=0.4.0
httpx>=0.27.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.4.22
